% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimFile_Creation_Functions.R
\name{BatchMixNovor}
\alias{BatchMixNovor}
\title{Wrapper of MixNovor() using a dataTracker row as input}
\usage{
BatchMixNovor(
  dataTracker_row,
  NovorResultList,
  export_csv = FALSE,
  outputFileDir = "."
)
}
\arguments{
\item{dataTracker_row}{a dataframe containing a single row. At minimum must
contain columns "ContributorDataset", "PrimaryDataset", "ContributorScans",
"SimID", and "SimDataset".}

\item{NovorResultList}{list of dataframes containing Novor results. The
column containing scan numbers must be called "ScanNum". PrimaryDataset and
ContributorDataset must match to entry names in NovorResultList.}

\item{export_csv}{Should the simulated raw Novor results be exported to a csv
file? Default FALSE}

\item{outputFileDir}{path to the desired output folder for raw Novor results.
Defaults to the current working directory. The filename is taken from the
SimDataset column with "_Raw_Novor.csv" appended.}
}
\value{
Returns a dataframe of Novor results consisting of all the primary
  results and the contributor results associated with the scan numbers in
  ContributorScans. A column containing the input SimID and SimDataset is also
  added.
}
\description{
\code{BatchMixNovor} is a wrapper of \code{\link{MixNovor}} which takes
parameters for \code{\link{MixNovor}} from dataTracker_row. This is necessary
so you can call \code{\link[plyr]{ddply}} on a dataTracker dataframe to make
batches of simulated Novor results. After calling \code{\link{MixNovor}} it
adds a column defining the simulated dataset name which will be needed to
make peptide list files. It gets this value from the SimDataset column in
dataTracker_row. You can use the \code{export_csv} and \code{outputFileDir}
parameters to export the simulated raw Novor results to a csv file in
addition to returning them as a dataframe. If you do not want the results
returned as a dataframe, call \code{BatchMixNovor} using
\code{\link[plyr]{d_ply}}.
}
\seealso{
Other Simulation Creation: 
\code{\link{BatchSelectScans}()},
\code{\link{MakeSimFile}()},
\code{\link{MixFilterDigest}()},
\code{\link{MixNovor}()},
\code{\link{SubsetMergeMzmls}()},
\code{\link{selectScans}()}

Other Novor Simulation: 
\code{\link{MixFilterDigest}()},
\code{\link{MixNovor}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Novor Simulation}
\concept{Simulation Creation}
