% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimFile_Creation_Functions.R
\name{BatchSelectScans}
\alias{BatchSelectScans}
\title{Select scans from MSGF+ results with SimRecipe input}
\usage{
BatchSelectScans(recipeRow, dbResults, rng_version = "R_version_default")
}
\arguments{
\item{recipeRow}{dataframe with a single row. At minimum must have columns:
"ContributorDataset", "SimID", "TargetNumContributorProteins", and
"TargetContributorPep_per_Prot". Any extra columns will be retained and
returned in output.}

\item{dbResults}{dataframe containing unrolled MSGF+ results (one protein per
row). Must have columns "cleanseq", "Protein", and "ScanNum"}

\item{rng_version}{R version string (Ex: '3.5.3') or 'R_version_default'
(default). The RNG (random number generator) state within selectScans() is
set to corrispond to the defaults of the R version specified in
\code{rng_version}. If 'R_version_default' then the default settings for the
currently installed R version are used.}
}
\value{
Returns the input recipeRow dataframe with added columns containing
  what and how many contributor scans, peptides, and proteins were added.
  Added scans, peptides, and proteins are stored as a list of
  character/integer vector.
}
\description{
\code{BatchSelectScans} uses parameters stored recipeRow and a dataframe of
MSGF+ results to call \code{\link{selectScans}}. It is meant to be used with
\code{\link[plyr]{ddply}} on a SimRecipes table and a list of contributor
MSGF+ results. The selected scans, and their associated peptides and
proteins, are added to the input recipeRow dataframe and returned.
}
\seealso{
Other Simulation Creation: 
\code{\link{BatchMixNovor}()},
\code{\link{MakeSimFile}()},
\code{\link{MixFilterDigest}()},
\code{\link{MixNovor}()},
\code{\link{SubsetMergeMzmls}()},
\code{\link{selectScans}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Simulation Creation}
