% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_raw_file.R
\name{convert_raw_file}
\alias{convert_raw_file}
\title{Convert raw files using ThermoRawFileParser}
\usage{
convert_raw_file(input_file, output_dir = NULL, output_format)
}
\arguments{
\item{input_file}{Path to the input RAW file.}

\item{output_dir}{Path to the desired output directory. If NULL (default) the
output will be in the same directory as input_file.}

\item{output_format}{Desired output format. Can be `mgf` or `mzML`.}
}
\value{
Returns the path to the output file.
}
\description{
This function is a wrapper for ThermoRawFileParser.exe which converts raw
files to mgf or mzML format. The mgf files it produces contain the `SCANS=`
line which allows Novor to include scan numbers in its output files.
ThermoRawFileParser.exe must be in the systems PATH such that it can run
regardless of its installation location.
}
\seealso{
\code{\link{runNovor}} ; \code{\link{mzMLtoMGF}}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
