% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataTracker_SimID5_8_mzMLex}
\alias{dataTracker_SimID5_8_mzMLex}
\title{Example dataTracker dataframe for simulation via mzML}
\format{
A dataframe with 4 rows and 21 variables: \describe{
  \item{ContributorDataset}{Filename w/o extension of the contributor mzML, chr}
  \item{SimID}{ID number of the simulated dataset, int}
  \item{PrimaryDataset}{Filename w/o extension of the primary mzML, chr}
  \item{PrimaryOrganism}{Name of the primary organism, chr}
  \item{ContributorOrganism}{Name of the contributor organism, chr}
  \item{Distance}{taxonomic distance between contributor and primary, num}
  \item{TargetNumContributorProteins}{desired number of proteins selected from
  contributor, int}
  \item{TargetContributorPep_per_Prot}{desired number of
  peptides/protein from contributor, int}
  \item{SimFileName}{Filename of the simulated dataset w/ extension, chr}
  \item{ContributorScans}{contributor scan numbers, vector of ints}
  \item{NumContributorScans}{total number of contributor scans, int}
  \item{ContributorPeptides}{contributor peptides, vector of chr}
  \item{NumContributorPeptides}{total number of contributor peptides, int}
  \item{ContributorProteins}{contributor proteins, vector or chr}
  \item{NumContributorProteins}{total number of contributor proteins, int}
  \item{Birthday}{date and time simulated dataset was created, POSIXct format
  yyyy-mm-dd hh:mm:ss}
  \item{MGF_name}{Filename of the MGF file used by
  Novor, chr}
  \item{Raw_Novor_Filename}{Filename of the Novor output file,
  chr}
  \item{Novor_ParamFile}{Filename of the Novor parameter file, chr}
  \item{PeptideListFile}{File containing a list of digested peptides created
  by makePepFile from the filtered Novor results, chr}
  \item{Num_Digested_Peptides}{number of digested peptides contained in the
  PeptideListFile, int} }
}
\source{
Result from running the mzML simulation workflow using the SimRecipes_SimID5_8 SimRecipes table described in the 'Simulation workflow using simulated mzMLs' vignette.
}
\usage{
dataTracker_SimID5_8_mzMLex
}
\description{
An example dataTracker dataframe which results from the mzML version of the
simulated dataset creation pipeline run using SimRecipes_SimID5_8.csv
}
\keyword{datasets}
