% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peptide_Process_Output_Functions.R
\name{digestPeptides}
\alias{digestPeptides}
\title{Digest a set of peptides and discard short ones}
\usage{
digestPeptides(peptides, prolineBlock = TRUE, discardLessThan = 6)
}
\arguments{
\item{peptides}{a character vector containing peptides}

\item{prolineBlock}{should P after R or K prevent splitting at that site,
defaults to TRUE}

\item{discardLessThan}{minimum length of output peptides, defaults to 6}
}
\value{
a character vector of peptides
}
\description{
\code{digestPeptides} returns set of peptides with no missed trypsin cut sites
created by digesting the input peptides (splitting at R or K) and discarding
those with a length less than 'discardLessThan'.
}
\seealso{
\code{\link{makePepFile}}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
