% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NovorFunctions.R
\name{filterNovor}
\alias{filterNovor}
\title{Filter Novor results using ppm error, Novor score, and peptide length}
\usage{
filterNovor(
  rawdata,
  ppmErrorMin = -15,
  ppmErrorMax = 15,
  minNovorScore = 70,
  minLength = 6
)
}
\arguments{
\item{rawdata}{dataframe of novor results of the type returned by
\code{importSingleNovor}, at minimum must have columns "cleanseq",
"NovorScore", and "ppmError"}

\item{ppmErrorMin}{minimum allowed precursor error, default -15}

\item{ppmErrorMax}{maximum allowed precursor error, default +15}

\item{minNovorScore}{minimum allowed Novor score, default 70}

\item{minLength}{minimum allowed peptide length, default 6}
}
\value{
Returns a dataframe containing rows that pass the filter parameters
}
\description{
This is a helper function which filters raw Novor results by user specified
cutoffs of ppm error, Novor Score, and peptide length. Peptide length is
calculated within the function based on the "cleanseq" column and added as
"length" column.
}
\seealso{
Other Novor Functions: 
\code{\link{SingleMGFIDtoScan}()},
\code{\link{importFilter_Novor}()},
\code{\link{importSingleNovor}()},
\code{\link{plotNovorHexBin}()},
\code{\link{runNovorSingle}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Novor Functions}
