% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NovorFunctions.R
\name{GG_save_pdf}
\alias{GG_save_pdf}
\title{Save list of ggplot2 objects to single pdf}
\usage{
GG_save_pdf(list, filename)
}
\arguments{
\item{list}{(list) List of ggplot2 objects.}

\item{filename}{(chr) What to call the pdf.}
}
\value{
Invisible NULL.
}
\description{
Got this from StackOverflow \url{https://stackoverflow.com/questions/12234248/printing-multiple-ggplots-into-a-single-pdf-multiple-plots-per-page}
}
