% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NovorFunctions.R
\name{importFilter_Novor}
\alias{importFilter_Novor}
\title{Import/Filter Novor results using dataframe style input}
\usage{
importFilter_Novor(
  trackingDF_row,
  ppmErrorMin = -15,
  ppmErrorMax = 15,
  minNovorScore = 70,
  minLength = 6
)
}
\arguments{
\item{trackingDF_row}{At minimum must have columns "Raw_Novor_Filename" and
"MGF_name"}

\item{ppmErrorMin}{minimum allowed precursor error, default -15}

\item{ppmErrorMax}{maximum allowed precursor error, default +15}

\item{minNovorScore}{minimum allowed Novor score, default 70}

\item{minLength}{minimum allowed peptide length, default 6}
}
\value{
Returns a dataframe of filtered Novor results with added metadata
}
\description{
\code{\link{importFilter_Novor}} calls \code{\link{importSingleNovor}} and
\code{\link{filterNovor}} using a single-row dataframe of the style created
by \code{\link{MakeSimFile}} and \code{\link{runNovorSingle}}. As such it is
meant to be used with dlply or \code{\link[plyr]{ddply}} on a dataframe
containing input parameters. It imports the Novor results, adds metadata from
the MGF file, then filters the results using user specifed parameters.
}
\seealso{
Other Novor Functions: 
\code{\link{SingleMGFIDtoScan}()},
\code{\link{filterNovor}()},
\code{\link{importSingleNovor}()},
\code{\link{plotNovorHexBin}()},
\code{\link{runNovorSingle}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Novor Functions}
