% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NovorFunctions.R
\name{importSingleNovor}
\alias{importSingleNovor}
\title{Import Novor results and add scan numbers}
\usage{
importSingleNovor(NovorOutputFile, MGF_file = NULL)
}
\arguments{
\item{NovorOutputFile}{path to a raw Novor output file. Can be missing the
metadata header but the column names are expected to be the same as in the
raw Novor output.}

\item{MGF_file}{path to the MGF file corresponding to the `NovorOutputFile`
(necessary if scan numbers not present in Novor output file). Default
`NULL`}
}
\value{
Returns a dataframe of unfiltered Novor results with scan numbers,
  clean peptide sequence, and names of the MGF and raw Novor output file
}
\description{
\code{importSingleNovor} imports Novor results, cleans up the column names,
adds a modification free peptide sequence (cleanseq). The function can
accommodate Novor files with or without the standard metadata header, however
the column names are expected to be the same as in the raw Novor output. If
the Novor output does not have scan numbers in the scanNum column (as
happens with MGF files made with MSConvert that have no `SCANS=` line`) then
\code{\link{SingleMGFIDtoScan}} is called to add correct scan numbers as well
as each scan's origin filename. It assumes modification info is always of the
pattern "(ModSymbol)".
}
\seealso{
Other Novor Functions: 
\code{\link{SingleMGFIDtoScan}()},
\code{\link{filterNovor}()},
\code{\link{importFilter_Novor}()},
\code{\link{plotNovorHexBin}()},
\code{\link{runNovorSingle}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Novor Functions}
