% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiProtein_Contributor_SimFile_Functions.R
\name{MakeMultiSimFile}
\alias{MakeMultiSimFile}
\title{MakeMultiSimFile
Creates a new mzML file that combines contains scans for
specific proteins from a database search, with a primary mzML file. This
function can take multiple proteins from multiple database searches, as long
as each unique combination of contributor file and protein has a different
row.}
\usage{
MakeMultiSimFile(multi_DT, MSGF_results, working_dir)
}
\arguments{
\item{multi_DT}{A dataframe which contains at least the following columns.
Dataset: A unique name for the final mzML file, without the extension;
SimID: Unique identifier; PrimaryDataset: Filename of the primary dataset
without the extension.; ContributorDataset: Filename of the contributor
dataset without the extension; ContributorProteins: Protein identifiers
that exactly match proteins in the MS-GF+ output.  Each row corresponds to
a unique combination of primary file, contributor file, and contributor
protein.}

\item{MSGF_results}{A dataframe of database search PSMs. The follow columns
are required: Dataset, Scan, and Protein. The Dataset column is expected to
have the exact same dataset name of the one_contrib_DT$ContributorDataset.
Additionally, The Protein column is expected to have the exact string that
is found in the one_contrib_DT$ContributorProteins column.}

\item{working_dir}{The full path to the mzML files and the location that the
new mzML files will be written to.}
}
\value{
Returns the input dataframe 'multi_DT' with two new columns. One is
  "SimFileName", which contains the name of the new mzML file created with
  the PrimaryDataset and the scans from the proteins from the
  ContributorDataset. The other is "Scans", which is a list of the scans
  selected for each protein from the MSGF_results dataframe.
}
\description{
MakeMultiSimFile
Creates a new mzML file that combines contains scans for
specific proteins from a database search, with a primary mzML file. This
function can take multiple proteins from multiple database searches, as long
as each unique combination of contributor file and protein has a different
row.
}
\examples{
\dontrun{
# In this example, the files "Sample_1.mzML", "Sample_2.mzML", "Sample_3.mzML",
# and "Primary_sample.mzML" are in the directory "C:/Users/XXX123/my_mzML_dir".
# The output file will be called "SimMulti_1.mzML" and it will also be in
# "C:/Users/XXX123/my_mzML_dir". The output file will have the scans that
# are in "filtered_MSGF_dataframe" that were identified in "Sample_1" as
# being from "FEL1B_FELCA", in "Sample_2" as being from "TRPV6_HUMAN", and
# in "Sample_3" as being from "Sample_3" along will all of the scans from
# "Primary_sample.mzML".
contrib_datasets <- c("Sample_1","Sample_2","Sample_3")
contrib_proteins <- c("FEL1B_FELCA", "TRPV6_HUMAN", "CTR2_HUMAN")
primary_dataset <- c("Primary_sample", "Primary_sample","Primary_sample")
example_multi_taxa_DT <- data.frame("Dataset"= c("SimMulti_1", "SimMulti_1", "SimMulti_1"),
                                    "SimID" = c(1, 1, 1),
                                    "PrimaryDataset" = primary_dataset,
                                    "ContributorDataset" = contrib_datasets,
                                    "ContributorProteins" = contrib_proteins,
                                    stringsAsFactors = FALSE)
mzML_dir <- "C:/Users/XXX123/my_mzML_dir"
load("filtered_MSGF_dataframe.RData")
example_multi_taxa_DT_results <- MakeMultiSimFile(example_multi_taxa_DT,
# filtered_MSGF_dataframe , mzML_dir)
}
}
\seealso{
Other Multi-contributor Simulations: 
\code{\link{MixMultipleNovor}()},
\code{\link{SingleDataset_SelectScans}()},
\code{\link{mergeMultMzML}()},
\code{\link{subsetMzMLProteinScans}()},
\code{\link{subsetSingleNovor}()}

Other Specific Protein Simulations: 
\code{\link{mergeMultMzML}()},
\code{\link{selectProteinScans}()},
\code{\link{subsetMzMLProteinScans}()}
}
\author{
Isabelle O'Bryon \email{isabelle.obryon@pnnl.gov}
}
\concept{Multi-contributor Simulations}
\concept{Specific Protein Simulations}
