% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peptide_Process_Output_Functions.R
\name{makePepFile}
\alias{makePepFile}
\title{Export digested peptides from a PSM dataframe}
\usage{
makePepFile(
  inputdf,
  datasetCol,
  outputFileDir = ".",
  sequenceCol = "cleanseq",
  prolineBlock = TRUE,
  discardLessThan = 6
)
}
\arguments{
\item{inputdf}{dataframe containing at least a column of peptide sequences
and a column containing the dataset name}

\item{datasetCol}{name of the column containing the dataset name, the first
value will be used as the export filename with "_peptides.csv" added}

\item{outputFileDir}{relative file path to the desired output folder, default
is the current working directory}

\item{sequenceCol}{name of the column containing the clean peptide sequences
(no modification info)}

\item{prolineBlock}{should P after R or K prevent splitting at that site,
defaults to TRUE}

\item{discardLessThan}{minimum length of output peptides, defaults to 6}
}
\value{
a dataframe row containing the name of the output file along with the
  number of peptides contained in the file
}
\description{
\code{makePepFile} takes a dataframe containing at least a column of peptide
sequences and a column containing the dataset name and writes a .csv file
containing the unique list of peptides resulting from digesting the peptides
with \code{\link{digestPeptides}}. The csv file contains a single column of
peptides with "x" as the first entry. The filename is the first entry in the
'datasetCol' column with "_peptides.csv" added. The function returns a
dataframe row containing the name of the output file along with the number of
peptides contained in the file. You can also select an output directory by
supplying a relative file path from the working directory.
}
\seealso{
\code{\link{digestPeptides}}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
