% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimFile_Creation_Functions.R
\name{MakeSimFile}
\alias{MakeSimFile}
\title{Make a simulated dataset from a SimRecipes row and MSGF+ results}
\usage{
MakeSimFile(recipeRow, dbResults, rng_version = "R_version_default")
}
\arguments{
\item{recipeRow}{dataframe with a single row. At minimum must have columns:
"ContributorDataset", "SimID", "PrimaryDataset", "PrimaryOrganism",
"ContributorOrganism", "TargetNumContributorProteins", and
"TargetContributorPep_per_Prot". Any extra columns will be retained and
returned in output.}

\item{dbResults}{dataframe containing unrolled MSGF+ results (one protein per
row). Must have columns "cleanseq", "Protein", and "ScanNum"}

\item{rng_version}{R version string (Ex: '3.5.3') or 'R_version_default'
(default). The RNG (random number generator) state within selectScans() is
set to corrispond to the defaults of the R version specified in
\code{rng_version}. If 'R_version_default' then the default settings for the
currently installed R version are used.}
}
\value{
Returns the input recipeRow dataframe with added columns containing
  what and how many contributor scans, peptides, and proteins were added.
  Added
  scans, peptides, and proteins are stored as a list of character/integer
  vectors.
}
\description{
\code{MakeSimFile} is meant to be used with ddply on a SimRecipes table and a
list of contributor MSGF+ results. It uses parameters stored the SimRecipes
table and a dataframe of MSGF+ results to call \code{\link{SubsetMergeMzmls}}
and \code{\link{selectScans}}. It keeps track of what scans, peptides, and
proteins were added from the contributor mzML along with when the simulated
dataset was made. This information is added to the input recipeRow dataframe
and returned. All input mzMLs must be in the working directory.
}
\seealso{
Other Simulation Creation: 
\code{\link{BatchMixNovor}()},
\code{\link{BatchSelectScans}()},
\code{\link{MixFilterDigest}()},
\code{\link{MixNovor}()},
\code{\link{SubsetMergeMzmls}()},
\code{\link{selectScans}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Simulation Creation}
