% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiProtein_Contributor_SimFile_Functions.R
\name{mergeMultMzML}
\alias{mergeMultMzML}
\title{mergeMultMzML
This function takes in a dataframe and will merge together the
primary mzML with the contributors mzML files.}
\usage{
mergeMultMzML(multi_contrib_DT)
}
\arguments{
\item{multi_contrib_DT}{A dataframe that contains the contributors, proteins
from the contributors, and the scans.}
}
\value{
Returns the input dataframe 'multi_contrib_DT' with a new column
  "SimFileName", which contains the name of the new mzML file created with
  the PrimaryDataset and the scans from the proteins from the
  ContributorDataset.
}
\description{
mergeMultMzML
This function takes in a dataframe and will merge together the
primary mzML with the contributors mzML files.
}
\examples{
\dontrun{
# This function is not meant to be called by the user, it is a function needed
# for the MakeMultiSimFile function.
multi_DT_scans <- ddply(multi_DT, .(ContributorDataset), selectProteinScans,  MSGF_results )
}
}
\seealso{
Other Multi-contributor Simulations: 
\code{\link{MakeMultiSimFile}()},
\code{\link{MixMultipleNovor}()},
\code{\link{SingleDataset_SelectScans}()},
\code{\link{subsetMzMLProteinScans}()},
\code{\link{subsetSingleNovor}()}

Other Specific Protein Simulations: 
\code{\link{MakeMultiSimFile}()},
\code{\link{selectProteinScans}()},
\code{\link{subsetMzMLProteinScans}()}
}
\author{
Isabelle O'Bryon \email{isabelle.obryon@pnnl.gov}
}
\concept{Multi-contributor Simulations}
\concept{Specific Protein Simulations}
