% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimFile_Creation_Functions.R
\name{MixFilterDigest}
\alias{MixFilterDigest}
\title{MixNovor, FiltNovor, makePepFile using dataTracker row}
\usage{
MixFilterDigest(
  dataTracker_row,
  NovorResultList,
  outputFileDir = ".",
  ppmErrorMin = -15,
  ppmErrorMax = 15,
  minNovorScore = 70,
  minLength = 6,
  datasetCol = "SimDataset",
  sequenceCol = "cleanseq",
  prolineBlock = TRUE,
  discardLessThan = 6
)
}
\arguments{
\item{dataTracker_row}{a dataframe containing a single row of the format
returned by \code{\link{BatchSelectScans}}. At minimum must contain columns
"ContributorDataset", "PrimaryDataset", "ContributorScans", "SimID", and
"SimDataset".}

\item{NovorResultList}{list of dataframes containing Novor results. The column
containing scan numbers must be called "ScanNum". PrimaryDataset and
ContributorDataset must match to entry names in NovorResultList.}

\item{outputFileDir}{relative file path to the desired output folder, default
is the current working directory}

\item{ppmErrorMin}{minimum allowed precursor error, default -15}

\item{ppmErrorMax}{maximum allowed precursor error, default +15}

\item{minNovorScore}{minimum allowed Novor score, default 70}

\item{minLength}{minimum allowed peptide length, default 6}

\item{datasetCol}{name of the column containing the dataset name, the first
value will be used as the export filename with "_peptides.csv" added}

\item{sequenceCol}{name of the column containing the clean peptide sequences
(no modification info)}

\item{prolineBlock}{should P after R or K prevent splitting at that site,
defaults to TRUE}

\item{discardLessThan}{minimum length of output peptides, defaults to 6}
}
\value{
Returns dataTracker_row with new columns containing the name of the
 peptide list file (PeptideListFile) and the number of digested peptides
 (Num_Digested_Peptides)
}
\description{
\code{MixFilterDigest} is meant to be used with large sets of simulated
datasets (> 50) to avoid holding more than 50 sets of simulated Novor results
in memory. To do this it calls \code{\link{MixNovor}},
\code{\link{filterNovor}}, and \code{\link{makePepFile}} givin a single
dataTracker row. This bypasses \code{\link{BatchMixNovor}}. The downside of
using this function is that you can't interrogate the the raw and filtered
mixed novor results.
}
\seealso{
Other Simulation Creation: 
\code{\link{BatchMixNovor}()},
\code{\link{BatchSelectScans}()},
\code{\link{MakeSimFile}()},
\code{\link{MixNovor}()},
\code{\link{SubsetMergeMzmls}()},
\code{\link{selectScans}()}

Other Novor Simulation: 
\code{\link{BatchMixNovor}()},
\code{\link{MixNovor}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Novor Simulation}
\concept{Simulation Creation}
