% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiContributor_SimFile_Functions.R
\name{MixMultipleNovor}
\alias{MixMultipleNovor}
\title{Simulated Novor results from multiple datasets}
\usage{
MixMultipleNovor(dataTracker_rows, RawNovorList)
}
\arguments{
\item{dataTracker_rows}{subset of rows from a dataTracker dataframe which
correspond to the component datasets of a single simulation}

\item{RawNovorList}{list of dataframes with each dataframe containing the
Novor results for a particular dataset. The name of each entry in the list
be the dataset name. The dataframe must have a column containing scan
numbers which must be called "ScanNum".}
}
\value{
a dataframe of simulated Novor results created from the datasets and
  scans listed in each row in dataTracker_row
}
\description{
\code{MixMultipleNovor} takes the subset of a dataTracker dataframe which
corresponds to a single simulation and calls \code{\link{subsetSingleNovor}}
using adply on each row and merges the results to create the simulated
dataframe of Novor results. It then creates and adds a simulated dataset name
as a new column.
}
\seealso{
Other Multi-contributor Simulations: 
\code{\link{MakeMultiSimFile}()},
\code{\link{SingleDataset_SelectScans}()},
\code{\link{mergeMultMzML}()},
\code{\link{subsetMzMLProteinScans}()},
\code{\link{subsetSingleNovor}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Multi-contributor Simulations}
