% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimFile_Creation_Functions.R
\name{MixNovor}
\alias{MixNovor}
\title{Merge contributor and primary Novor results}
\usage{
MixNovor(
  PrimaryDataset,
  ContributorDataset,
  NovorResultList,
  ContributorScans,
  SimID
)
}
\arguments{
\item{PrimaryDataset}{name of the primary dataset}

\item{ContributorDataset}{name of the contributor dataset}

\item{NovorResultList}{list of dataframes containing Novor results. The
column containing scan numbers must be called "ScanNum". PrimaryDataset and
ContributorDataset must match to entry names in NovorResultList.}

\item{ContributorScans}{vector of scan numbers to take from the contributor
results and add to the primary results}

\item{SimID}{Simulated dataset ID, will be added as a column to the output}
}
\value{
Returns a dataframe of Novor results consisting of all the primary
  results and the contributor results associated with the scan numbers in
  ContributorScans. A column containing the input SimID is also added.
}
\description{
\code{MixNovor} extracts the Novor results corrisponding to the scan numbers
in ContributorScans and adds them to the Novor results of the primary
dataset.
\strong{Note:}To avoid duplication of scan numbers, it is reccomended to
have a column in the input Novor result dataframes which lists the origin
file. This is added if you use \code{\link{importSingleNovor}} to import the
Novor results from the csv files.
}
\seealso{
Other Simulation Creation: 
\code{\link{BatchMixNovor}()},
\code{\link{BatchSelectScans}()},
\code{\link{MakeSimFile}()},
\code{\link{MixFilterDigest}()},
\code{\link{SubsetMergeMzmls}()},
\code{\link{selectScans}()}

Other Novor Simulation: 
\code{\link{BatchMixNovor}()},
\code{\link{MixFilterDigest}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Novor Simulation}
\concept{Simulation Creation}
