% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiDataTracker_SimID5_8}
\alias{multiDataTracker_SimID5_8}
\title{Example dataTracker dataframe from multi-organism workflow}
\format{
A data frame with 4 rows and 8 variables: \describe{
  \item{SimID}{Numeric identifier unique to each simulation, int}
  \item{Dataset}{Name of a dataset, has no extension or "_Novor" suffix , chr}
  \item{Type}{Either "Primary" or "Contributor", chr}
  \item{Organism}{name of the organism associated with the dataset, chr}
  \item{TargetNumContributorProteins}{How many proteins to add from the
  contributor dataset  (This should be "NA" for Primary datasets), int}
  \item{TargetContributorPep_per_Prot}{How many peptides per protein to add
  from each selected contributor protein (This should be "NA" for Primary
  datasets), int}
  \item{Project}{Name of the project associated with the simulation which is
  used to name the output peptide list files, chr}
  \item{ContributorScans}{list of scan numbers selected from the dataset (NA if a primary dataset), list of ints}
  \item{NumContributorScans}{Number of scans selected from the dataset (NA is a primary dataset), int}
  \item{ContributorPeptides}{list of peptides associated with the selected scans (NA if a primary dataset), list of character strings}
  \item{NumContributorPeptides}{number of unique peptides assocated with the selected scans (NA if a primary dataset), int}
  \item{ContributorProteins}{list of proteins associated with the selected scans (NA if a primary dataset), list of character strings}
  \item{NumContributorProteins}{number of unique proteins assocated with the selected scans (NA if a primary dataset), int}
  \item{PeptideListFile}{name of the file containing the list of digested peptides for each simulation, chr}
  \item{Num_Digested_Peptides}{number of digested peptides for each simulation , duplicated for each row associated with a simulation, int}
  }
}
\source{
Multi-dataset simulation workflow run on SimRecipes5_8_long
}
\usage{
multiDataTracker_SimID5_8
}
\description{
This is an example of the dataTracker dataframe which would result from
running the multi-organism workflow described in the "Make simulations from
many organisms using Novor results" vignette using SimRecipes5_8_long as the
input SimRecipes table.
}
\keyword{datasets}
