% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NovorFunctions.R
\name{mzMLtoMGF}
\alias{mzMLtoMGF}
\title{Convert mzML to MGF using MSconvert and a config file}
\usage{
mzMLtoMGF(mzMLname, outputDir = ".", configFile)
}
\arguments{
\item{mzMLname}{path to the mzML file (with mzML extension)}

\item{outputDir}{path to the desired output directory, defaults to the current
working directory.}

\item{configFile}{MSConvert config file containing MGF scan title format. A
default file is included with the package which can be accessed using
\code{system.file("extdata","convertMGF_config.txt",
package="MakeSearchSim", mustWork = TRUE)}.}
}
\value{
Returns the path to the mgf file which will be writtin in the
 \code{outputDir} directory.
}
\description{
\code{mzMLtoMGF} is a wrapper for a command line call to MSConvert which
converts the input mzML file to an MGF file with scan titles described in the
configFile. \code{msconvert.exe} must be in you PATH.
}
\seealso{
\code{\link{convert_raw_file}}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
