% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NovorFunctions.R
\name{plotNovorHexBin}
\alias{plotNovorHexBin}
\title{Precursor error by Novor score hexbin plot}
\usage{
plotNovorHexBin(
  NovorDF,
  dataset_column = "MGFDataset",
  ppmMin = -15,
  ppmMax = 15,
  scoreMin = 20
)
}
\arguments{
\item{NovorDF}{a dataframe of Novor PSMs. Must have columns: "ppmError" and
"NovorScore", if not specified dataset name assumed to be in "MGFDataset"
column.}

\item{dataset_column}{name of the column containing the dataset name, first
entry will be used in the graph title.}

\item{ppmMin}{Minimum precursor error to plot}

\item{ppmMax}{Maximum precursor error to plot}

\item{scoreMin}{Minimum Novor score to plot}
}
\value{
Returns a ggplot2 plot object
}
\description{
\code{plotNovorHexBin} makes a hexbin plot of precursor error by Novor score
after filtering using the input parameters. Title is made using first value
in MGFDataset column.
}
\seealso{
Other Novor Functions: 
\code{\link{SingleMGFIDtoScan}()},
\code{\link{filterNovor}()},
\code{\link{importFilter_Novor}()},
\code{\link{importSingleNovor}()},
\code{\link{runNovorSingle}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Novor Functions}
