% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NovorFunctions.R
\name{runNovor}
\alias{runNovor}
\title{Run Novor given an MGF and parameter file}
\usage{
runNovor(MGF_File, NovorParams, Novor_Path)
}
\arguments{
\item{MGF_File}{name of an MGF file in the working directory}

\item{NovorParams}{path to a Novor parameter file}

\item{Novor_Path}{Windows style path to the Novor .bat file within the Novor
installaion folder. Backslashes must be escaped by a second backslash. Ex:
`C:\\\\Novor\\\\win\\\\novor.bat`}
}
\value{
Returns the name of the Novor result file
}
\description{
\code{runNovor} is a wrapper function to run Novor on the command line. It
takes an MGF and Novor parameter file in the current working directory, runs
Novor, and returns the name of the Novor result file. The Novor result file
is the name of the input MGF file (minus extension) with "_Novor.csv" added as
a suffix.
}
\details{
Two Novor parameter files are included in the package and can be accessed
using \code{system.file("extdata",
"NovorParams_fTol002_pTol15_MetOx_noAlk_Novor_params.txt",
package="MakeSearchSim", mustWork = TRUE)}.

\code{NovorParams_fTol002_pTol15_MetOx_Alk_Novor_params.txt} has a static
carbamidomethylation modification and a variable methionine oxidation
modification. \code{NovorParams_fTol002_pTol15_MetOx_noAlk_Novor_params.txt}
only contains a variable methionine oxidation modification.
}
\examples{
\dontrun{
runNovor(MGF_File = "test.mgf",
        NovorParams = system.file("extdata",
             "NovorParams_fTol002_pTol15_MetOx_noAlk_Novor_params.txt",
              package="MakeSearchSim", mustWork = TRUE),
        Novor_Path = "C:\\\\Novor\\\\win\\\\novor.bat")
}

}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Novor functions}
