% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NovorFunctions.R
\name{runNovorSingle}
\alias{runNovorSingle}
\title{Convert mzML and run Novor using a dataTracker row}
\usage{
runNovorSingle(
  trackingDF_row,
  msConvert_config,
  novor_config,
  Novor_Path,
  mgf_output_dir = "."
)
}
\arguments{
\item{trackingDF_row}{dataframe with a single row. Must contain columns
"SimFileName" and}

\item{msConvert_config}{name of msConvert config file in the working
directory}

\item{novor_config}{name of the Novor parameter file in the working directory}

\item{Novor_Path}{Windows style path to the Novor .bat file within the Novor
installaion folder. Backslashes must be escaped by a second backslash. Ex:
`C:\\\\Novor\\\\win\\\\novor.bat`}

\item{mgf_output_dir}{path to the directory where MGF files should be written
by \code{\link{mzMLtoMGF}} . Defaults to the current working directory.}
}
\value{
Returns a dataframe with a single row. It is the input trackingDF_row
  with extra columns for the MGF filename, the Novor output filename, and the
  Novor parameter filename
}
\description{
\code{runNovorSingle} calls \code{\link{mzMLtoMGF}} then
\code{\link{runNovor}} using the mzML file in the "SimFileName" column of
trackingDF_row and the msConvert and Novor parameter files.
}
\seealso{
\code{\link{runNovor}} ; \code{\link{mzMLtoMGF}}

Other Novor Functions: 
\code{\link{SingleMGFIDtoScan}()},
\code{\link{filterNovor}()},
\code{\link{importFilter_Novor}()},
\code{\link{importSingleNovor}()},
\code{\link{plotNovorHexBin}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Novor Functions}
