% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiProtein_Contributor_SimFile_Functions.R
\name{selectProteinScans}
\alias{selectProteinScans}
\title{selectProteinScans
This function is called by MakeMultiSimFile to select all
of the scans from one dataset for specific proteins.}
\usage{
selectProteinScans(one_contrib_DT, dbResults)
}
\arguments{
\item{one_contrib_DT}{A dataframe that must have the columns:
ContributorDataset, ContributorProteins. It is expected that there will
only be one unique ContributorDataset. There may be one or more
ContributorProteins. DT stands for data tracker.}

\item{dbResults}{A dataframe of database search PSMs. The follow columns are
required: Dataset, Scan, and Protein. The Dataset column is expected to
have the exact same dataset name of the one_contrib_DT$ContributorDataset.
Additionally, The Protein column is expected to have the exact string that
is found in the one_contrib_DT$ContributorProteins column.}
}
\value{
A dataframe that is the same as the input dataframe one_contrib_DT
  with the added column "Scans". The column "Scans" is a list of the scans
  from the database result dataframe that had the given protein identifier in
  the Protein column.
}
\description{
selectProteinScans
This function is called by MakeMultiSimFile to select all
of the scans from one dataset for specific proteins.
}
\examples{
\dontrun{
# This function is not meant to be called by the user, it is a function
# needed for the MakeMultiSimFile function.
# load("filtered_MSGF_dataframe.RData")
merge_sample_1_w_2 <- data.frame("Dataset"=  "SimMulti_1",
                                    "SimID" = 1,
                                    "PrimaryDataset" = "Sample_1",
                                    "ContributorDataset" = "Sample_2",
                                    "ContributorProteins" = "FEL1B_FELCA",
                                    stringsAsFactors = FALSE)
merge_sample_1_w_2_Scans <- selectProteinScans(merge_sample_1_w_2, filtered_MSGF_dataframe)
}
}
\seealso{
Other Specific Protein Simulations: 
\code{\link{MakeMultiSimFile}()},
\code{\link{mergeMultMzML}()},
\code{\link{subsetMzMLProteinScans}()}
}
\author{
Isabelle O'Bryon \email{isabelle.obryon@pnnl.gov}
}
\concept{Specific Protein Simulations}
