% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimFile_Creation_Functions.R
\name{selectScans}
\alias{selectScans}
\title{Randomly select scans representing n proteins and m peptides}
\usage{
selectScans(
  databaseResults,
  Nprot,
  Mpep,
  SimID,
  rng_version = "R_version_default"
)
}
\arguments{
\item{databaseResults}{a dataframe containing unrolled MSGF+ results (one
protein per row). Must have columns "cleanseq", "Protein", and "ScanNum"}

\item{Nprot}{how many proteins to select. Note:  It is possible to end up with
more contributor proteins represented in the output scans if a selected
peptide belongs to more than one protein.}

\item{Mpep}{how many peptides per protein to select}

\item{SimID}{value used as the seed of \code{\link[base]{sample}} when
selecting peptides and proteins. Set as the ID of the simulated dataset when
called by \code{\link{MakeSimFile}}}

\item{rng_version}{R version string (Ex: '3.5.3') or 'R_version_default'
(default). The RNG (random number generator) state within selectScans() is
set to corrispond to the defaults of the R version specified in
\code{rng_version}. If 'R_version_default' then the default settings for the
currently installed R version are used.}
}
\value{
Returns an integer vector containing all scan numbers associated with
 the selected peptides and proteins
}
\description{
\code{selectScans} returns a set of scan numbers associated with a randomly
selected number of proteins(Nprot) and randomly selected number of
peptides(Mpep) per protein. SimID is used as the seed for the random
selection.
}
\details{
This funciton first calculates the number of peptides per protein and filters
out all proteins with < 2 peptides per protein. Of the remaining proteins,
"Nprot" are randomly selected. For each selected protein, a random set of
"Mpep" peptides are selected. All scans associated with those peptides are
then returned. If a protein has fewer peptides than Mpep, then all peptides
from that protein are used. All scans associated with each selected peptide
are returned. \strong{Warning:} This function only works with "unrolled" database
results, meaning one row for each scan, peptide, and protein combination.
}
\section{Random Number Generator (RNG) State}{

 The random number generation settings are a
 \strong{global variable}. Within \code{selectScans} they are changed
 according to the \code{rng_verison} parameter then the original state is
 restored before the function exits. Version 3.6.0 added a new setting
 \code{sample.kind} with values 'Rounding'(< 3.6.0) and 'Rejection'(3.6.0
 behavior). See \code{\link[base]{Random}} for more details.
}

\seealso{
Other Simulation Creation: 
\code{\link{BatchMixNovor}()},
\code{\link{BatchSelectScans}()},
\code{\link{MakeSimFile}()},
\code{\link{MixFilterDigest}()},
\code{\link{MixNovor}()},
\code{\link{SubsetMergeMzmls}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}; Natalie Heller
}
\concept{Simulation Creation}
