% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimRecipes5_8_long}
\alias{SimRecipes5_8_long}
\title{Example long format (multi-organism) SimRecipes table for SimID 5-8}
\format{
A data frame with 4 rows and 8 variables: \describe{
  \item{SimID}{Numeric identifier unique to each simulation, int}
  \item{Dataset}{Name of a dataset, has no extension or "_Novor" suffix , chr}
  \item{Type}{Either "Primary" or "Contributor", chr}
  \item{Organism}{name of the organism associated with the dataset, chr}
  \item{TargetNumContributorProteins}{How many proteins to add from the
  contributor dataset  (This should be "NA" for Primary datasets), int}
  \item{TargetContributorPep_per_Prot}{How many peptides per protein to add
  from each selected contributor protein (This should be "NA" for Primary
  datasets), int}
  \item{Project}{Name of the project associated with the simulation which is
  used to name the output peptide list files, chr}
  }
}
\source{
Example input included in package
}
\usage{
SimRecipes5_8_long
}
\description{
This is an example of a long format SimRecipes table used as input to the
multi-organism simulation pipeline described in the "Make simulations from
many organisms using Novor results" vignette. It describes the same
simulations as SimRecipes_SimID5_8.
}
\keyword{datasets}
