% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimRecipes_SimID5_8}
\alias{SimRecipes_SimID5_8}
\title{Example dual organism SimRecipes table for SimID 5-8}
\format{
A data frame with 4 rows and 8 variables: \describe{
  \item{ContributorDataset}{Filename (w/o extension) of the contributor mzML, chr}
  \item{SimID}{ID number of the simulated dataset, int}
  \item{PrimaryDataset}{Filename (w/o extension) of the primary mzML, chr}
  \item{PrimaryOrganism}{Name of the primary organism, chr}
  \item{ContributorOrganism}{Name of the contributor organism, chr}
  \item{Distance}{taxonomic distance between contributor and primary, num}
  \item{TargetNumContributorProteins}{desired number of proteins selected from
  contributor, int}
  \item{TargetContributorPep_per_Prot}{desired number of
  peptides/protein from contributor, int}
  \item{SimFileName}{Name chosen for the simulated dataset described by each row}
  }
}
\source{
Example included in package
}
\usage{
SimRecipes_SimID5_8
}
\description{
This is an example of a SimRecipes table used as input to the workflows which
only make simulations from a single primary and contributor dataset. It is
used in the 'Simulation workflow using simulated mzMLs' and 'Make simulated
peptide lists by mixing prime and contributor Novor results' vignettes.
}
\keyword{datasets}
