% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiContributor_SimFile_Functions.R
\name{SingleDataset_SelectScans}
\alias{SingleDataset_SelectScans}
\title{Select scans from a long format SimRecipes row}
\usage{
SingleDataset_SelectScans(
  recipeRow,
  dbResultList,
  rng_version = "R_version_default"
)
}
\arguments{
\item{recipeRow}{single row of a long format SimRecipes table (as a
dataframe) must contain columns "SimID", "Type", "Dataset",
"TargetNumContributorProteins", "TargetContributorPep_per_Prot". All other
columns are passed through to output.}

\item{dbResultList}{list of dataframes containing unrolled MSGF+ results (one
protein per row) one of which must contain the dataset referenced in
recipeRows. The dataset must match to the name of one of the entries in the
list. dataframe must have columns "cleanseq", "Protein", and "ScanNum"}

\item{rng_version}{R version string (Ex: '3.5.3') or 'R_version_default'
(default). The RNG (random number generator) state within selectScans() is
set to corrispond to the defaults of the R version specified in
\code{rng_version}. If 'R_version_default' then the default settings for the
currently installed R version are used.}
}
\value{
Returns the input row with added columns containing selected scans
  and what peptides and proteins they corrispond to.
}
\description{
\code{SingleDataset_SelectScans} uses parameters stored in the input
\code{recipeRow} and uses them to isolate the correct set of MSGF+ results
from \code{dbResultList} and call \code{\link{selectScans}}. It then adds
columns for what and how many scans, peptides, and proteins were selected
from the MSGF+ results.
}
\seealso{
Other Multi-contributor Simulations: 
\code{\link{MakeMultiSimFile}()},
\code{\link{MixMultipleNovor}()},
\code{\link{mergeMultMzML}()},
\code{\link{subsetMzMLProteinScans}()},
\code{\link{subsetSingleNovor}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Multi-contributor Simulations}
