% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NovorFunctions.R
\name{SingleMGFIDtoScan}
\alias{SingleMGFIDtoScan}
\title{Extract position, scan number, and original filename from an MGF file}
\usage{
SingleMGFIDtoScan(filepath, chunkSize = 1e+06)
}
\arguments{
\item{filepath}{path to an MGF file}

\item{chunkSize}{number of lines to read from MGF at a time}
}
\value{
Returns a dataframe with columns representing the scan position
 (MGForder), scan number (ScanNum), Original dataset (OriginDataset), and
 input filename w/o extension (MGFDataset)
}
\description{
\code{SingleMGFIDtoScan} reads an MGF file and returns a dataframe which
correlates the position of the scan in the MGF file to the scan number and
original file name contained in each title line. It expects the scan titles to
be in the form made by MSConvert using this filter: filter="titleMaker
<RunId>.<ScanNumber>.<ScanNumber>.<ChargeState> File:"<SourcePath>",
NativeID:"<Id>""
}
\details{
This function is optimized for speed at the cost of memory by reading the mgf
file into memory all at once. Reading line by line ends up being either the
same speed or slower, I think it's because read/write speed is limited by the
hard drive.
}
\seealso{
Other Novor Functions: 
\code{\link{filterNovor}()},
\code{\link{importFilter_Novor}()},
\code{\link{importSingleNovor}()},
\code{\link{plotNovorHexBin}()},
\code{\link{runNovorSingle}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}; Anthony Valente
}
\concept{Novor Functions}
