% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimFile_Creation_Functions.R
\name{SubsetMergeMzmls}
\alias{SubsetMergeMzmls}
\title{Subset a contributor mzML and merge with a primary mzML}
\usage{
SubsetMergeMzmls(scans, contributorName, primaryName, simname, simID)
}
\arguments{
\item{scans}{a integer vector containing scan numbers to select from the
contributor}

\item{contributorName}{dataset name of the contributor mzML file (no ".mzML"
extension)}

\item{primaryName}{dataset name of the primary mzML (no ".mzML" extension)}

\item{simname}{filename of the simulated dataset (\strong{with} the ".mzML"
extension)}

\item{simID}{ID number of the simulated dataset. Used to name the mzML
containing the subset of scans from the contributor mzML and passed to
\code{\link{selectScans}}}
}
\value{
Returns the name of the simulated dataset that was created
}
\description{
\code{SubsetMergeMzmls} first makes a new mzML conatining a specified set of
scans from the contributor mzML using the MSConvert filter parameter. It then
merges the subset contributor mzML with the primary mzML to make a new
simulated mzML with a specified file name. The contributor and primary mzMLs
must be in the working directory. MSConvert must be installed and in the
computer's PATH.
}
\seealso{
Other Simulation Creation: 
\code{\link{BatchMixNovor}()},
\code{\link{BatchSelectScans}()},
\code{\link{MakeSimFile}()},
\code{\link{MixFilterDigest}()},
\code{\link{MixNovor}()},
\code{\link{selectScans}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Simulation Creation}
