% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiProtein_Contributor_SimFile_Functions.R
\name{subsetMzMLProteinScans}
\alias{subsetMzMLProteinScans}
\title{subsetMzMLProteinScans
This function takes in a dataframe that has one unqiue
'ContributorDataset'. The scans listed in the 'Scans' column are selected
from the Contributor mzML file and written to a separate mzML file.}
\usage{
subsetMzMLProteinScans(one_contrib_DT)
}
\arguments{
\item{one_contrib_DT}{A dataframe that contains the contributors, proteins
from the contributors, and the scans}
}
\value{
The name of the new mzML file
}
\description{
subsetMzMLProteinScans
This function takes in a dataframe that has one unqiue
'ContributorDataset'. The scans listed in the 'Scans' column are selected
from the Contributor mzML file and written to a separate mzML file.
}
\examples{
\dontrun{
# This function is not meant to be called by the user, it is a function needed
# for the MakeMultiSimFile function.
subset_mzML_files <- ddply(multi_contrib_DT, .(ContributorDataset), subsetMzMLProteinScans )
}
}
\seealso{
Other Multi-contributor Simulations: 
\code{\link{MakeMultiSimFile}()},
\code{\link{MixMultipleNovor}()},
\code{\link{SingleDataset_SelectScans}()},
\code{\link{mergeMultMzML}()},
\code{\link{subsetSingleNovor}()}

Other Specific Protein Simulations: 
\code{\link{MakeMultiSimFile}()},
\code{\link{mergeMultMzML}()},
\code{\link{selectProteinScans}()}
}
\author{
Isabelle O'Bryon \email{isabelle.obryon@pnnl.gov}
}
\concept{Multi-contributor Simulations}
\concept{Specific Protein Simulations}
