% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiContributor_SimFile_Functions.R
\name{subsetSingleNovor}
\alias{subsetSingleNovor}
\title{Use long format dataTracker row to get Novor results}
\usage{
subsetSingleNovor(dataTracker_row, RawNovorList)
}
\arguments{
\item{dataTracker_row}{a row of a dataTracker dataframe produced by
\code{\link{SingleDataset_SelectScans}} must contain at least columns
"SimID", "Dataset", "Type", "ContributorScans"}

\item{RawNovorList}{list of dataframes with each dataframe containing the
Novor results for a particular dataset. The name of each entry in the list
be the dataset name. The dataframe must have a column containing scan
numbers which must be called "ScanNum".}
}
\value{
A dataframe containing the appropriate set of Novor results for the
  dataset referenced in dataTracker_row
}
\description{
\code{subsetSingleNovor} takes single row of a long format dataTracker from
the multiple dataset / simulation pipeline and returns the appropriate set of
Novor results. It returns all the results for the dataset if Type = Primary
and only the subset according to ContributorScans if the Type = Contributor.
Adds a column for SimID and for the dataset of origin(OriginDataset) to the
exported dataframe.
}
\seealso{
Other Multi-contributor Simulations: 
\code{\link{MakeMultiSimFile}()},
\code{\link{MixMultipleNovor}()},
\code{\link{SingleDataset_SelectScans}()},
\code{\link{mergeMultMzML}()},
\code{\link{subsetMzMLProteinScans}()}
}
\author{
Sarah C. Jenson \email{sarah.jenson@pnnl.gov}
}
\concept{Multi-contributor Simulations}
